@@ -9,13 +9,13 @@
   },
   "private": true,
   "engines": {
-    "node": "^18",
+    "node": "^20",
     "yarn": "^4"
   },
   "type": "module",
   "scripts": {
     "js": "rollup -c",
-    "sass": "node scripts/compile-sass.cjs",
+    "sass": "node scripts/compile-sass.js",
     "sassdoc": "sassdoc src/scss/**/*.scss",
     "prettier:js": "prettier --write \"**/*.{js,cjs}\"",
     "lint:other": "prettier --write \"**/*.{json,yml,scss}\"",
