@@ -1,10 +1,12 @@
 export const collection = [
   {
-    inputPath: './test1.md',
-    fileSlug: 'test1',
-    outputPath: './_site/test1/index.html',
-    url: '/test1/',
-    date: '2018-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test1.md',
+      fileSlug: 'test1',
+      outputPath: './_site/test1/index.html',
+      url: '/test1/',
+      date: '2018-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       tags: ['tag 1', 'tag2'],
@@ -22,138 +24,158 @@ export const collection = [
         },
       ],
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test2.md',
-    fileSlug: 'test2',
-    outputPath: './_site/test2/index.html',
-    url: '/test2/',
-    date: '2018-03-09T04:10:17.000Z',
+    page: {
+      inputPath: './test2.md',
+      fileSlug: 'test2',
+      outputPath: './_site/test2/index.html',
+      url: '/test2/',
+      date: '2018-03-09T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       tags: ['tag 1', 'tag2'],
       date: 'Last Modified',
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test3.md',
-    fileSlug: 'test3',
-    outputPath: './_site/test3/index.html',
-    url: '/test3/',
-    date: '2018-03-09T04:10:17.000Z',
+    page: {
+      inputPath: './test3.md',
+      fileSlug: 'test3',
+      outputPath: './_site/test3/index.html',
+      url: '/test3/',
+      date: '2018-03-09T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       tags: ['tag 1', 'tag2'],
       date: '2020-04-10T04:10:17.000',
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test4.md',
-    fileSlug: 'test4',
-    outputPath: './_site/test4/index.html',
-    url: '/test4/',
-    date: '2020-04-10T04:10:17.000Z',
+    page: {
+      inputPath: './test4.md',
+      fileSlug: 'test4',
+      outputPath: './_site/test4/index.html',
+      url: '/test4/',
+      date: '2020-04-10T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       tags: ['tag 1', 'tag2'],
       date: '2020-04-10T04:10:17.000Z',
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test5.md',
-    fileSlug: 'test5',
-    outputPath: './_site/test5/index.html',
-    url: '/test5/',
-    date: '2020-04-10T04:10:17.000Z',
+    page: {
+      inputPath: './test5.md',
+      fileSlug: 'test5',
+      outputPath: './_site/test5/index.html',
+      url: '/test5/',
+      date: '2020-04-10T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       index: 'tag 1',
       date: '2020-04-10T04:10:17.000Z',
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
 ];
 
 export const collection2 = [
   {
-    inputPath: './test1.md',
-    fileSlug: 'test1',
-    outputPath: './_site/test1/index.html',
-    url: '/test1/',
-    date: '2018-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test1.md',
+      fileSlug: 'test1',
+      outputPath: './_site/test1/index.html',
+      url: '/test1/',
+      date: '2018-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       tags: ['tag 1', 'tag2'],
       author: 'miriam',
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test2.md',
-    fileSlug: 'test1',
-    outputPath: './_site/test2/index.html',
-    url: '/test2/',
-    date: '2018-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test2.md',
+      fileSlug: 'test1',
+      outputPath: './_site/test2/index.html',
+      url: '/test2/',
+      date: '2018-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       tags: ['tag 1', 'tag2'],
       author: 'oddbird',
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test3.md',
-    fileSlug: 'test3',
-    outputPath: './_site/test3/index.html',
-    url: '/test1/',
-    date: '2018-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test3.md',
+      fileSlug: 'test3',
+      outputPath: './_site/test3/index.html',
+      url: '/test1/',
+      date: '2018-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       tags: ['tag 1', 'tag2'],
       author: 'erica',
       bird: 'erica',
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test4.md',
-    fileSlug: 'test4',
-    outputPath: './_site/test4/index.html',
-    url: '/test1/',
-    date: '2018-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test4.md',
+      fileSlug: 'test4',
+      outputPath: './_site/test4/index.html',
+      url: '/test1/',
+      date: '2018-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       tags: ['tag 1', 'tag2'],
       author: 'jimbob',
       end: '2019-01-09T04:10:17.000Z',
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test5.md',
-    fileSlug: 'test5',
-    outputPath: './_site/test5/index.html',
-    url: '/test5/',
-    date: '2018-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test5.md',
+      fileSlug: 'test5',
+      outputPath: './_site/test5/index.html',
+      url: '/test5/',
+      date: '2018-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       tags: ['tag 1', 'tag2'],
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
 ];
 
 export const collection3 = [
   {
-    inputPath: './test1.md',
-    fileSlug: 'test1',
-    outputPath: './_site/test1/index.html',
-    url: '/test1/',
-    date: '2019-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test1.md',
+      fileSlug: 'test1',
+      outputPath: './_site/test1/index.html',
+      url: '/test1/',
+      date: '2019-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Test Title',
       tags: ['tag 1', 'tag2'],
@@ -161,42 +183,48 @@ export const collection3 = [
       slug: 'news',
       oss: 'owner',
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test2.md',
-    fileSlug: 'test2',
-    outputPath: './_site/test2/index.html',
-    url: '/test2/',
-    date: '2019-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test2.md',
+      fileSlug: 'test2',
+      outputPath: './_site/test2/index.html',
+      url: '/test2/',
+      date: '2019-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Draft Title',
       tags: ['tag3', 'tag3'],
       author: 'erica',
       draft: true,
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test3.md',
-    fileSlug: 'test3',
-    outputPath: './_site/test3/index.html',
-    url: '/test3/',
-    date: '2019-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test3.md',
+      fileSlug: 'test3',
+      outputPath: './_site/test3/index.html',
+      url: '/test3/',
+      date: '2019-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Draft Title',
       tags: ['tag3', 'tag3'],
       author: 'erica',
       end: '2020-01-09T04:10:17.000Z',
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test4.md',
-    fileSlug: 'test4',
-    outputPath: './_site/test4/index.html',
-    url: '/test4/',
-    date: '2019-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test4.md',
+      fileSlug: 'test4',
+      outputPath: './_site/test4/index.html',
+      url: '/test4/',
+      date: '2019-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Event Title',
       tags: ['tag3', 'tag3'],
@@ -217,42 +245,48 @@ export const collection3 = [
         },
       ],
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
 ];
 
 export const collection4 = [
   {
-    inputPath: './test1.md',
-    fileSlug: 'test1',
-    outputPath: './_site/test1/index.html',
-    url: '/test1/',
-    date: '2040-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test1.md',
+      fileSlug: 'test1',
+      outputPath: './_site/test1/index.html',
+      url: '/test1/',
+      date: '2040-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Draft Title',
       tags: ['Video', 'Article'],
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './test3.md',
-    fileSlug: 'test3',
-    outputPath: './_site/test3/index.html',
-    url: '/test3/',
-    date: '2018-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './test3.md',
+      fileSlug: 'test3',
+      outputPath: './_site/test3/index.html',
+      url: '/test3/',
+      date: '2018-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Draft Title',
       tags: ['Article'],
       home: false,
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
   {
-    inputPath: './events.md',
-    fileSlug: 'events',
-    outputPath: './_site/events/index.html',
-    url: '/events/',
-    date: '2011-01-09T04:10:17.000Z',
+    page: {
+      inputPath: './events.md',
+      fileSlug: 'events',
+      outputPath: './_site/events/index.html',
+      url: '/events/',
+      date: '2011-01-09T04:10:17.000Z',
+    },
     data: {
       title: 'Draft Title',
       tags: ['Link', 'Article'],
@@ -268,7 +302,7 @@ export const collection4 = [
         },
       ],
     },
-    templateContent: '<h1>This is my title</h1>\n\n<p>This is content…',
+    content: '<h1>This is my title</h1>\n\n<p>This is content…',
   },
 ];
 
